+++ craft_platforms/_build.py
@@ -17,7 +17,7 @@
 
 from typing import Any, Callable, Dict, Iterable
 
-from craft_platforms import charm, deb, rock, snap
+from craft_platforms import charm, deb, rock, snap, validators
 from craft_platforms._buildinfo import BuildInfo
 from craft_platforms._platforms import get_platforms_build_plan
 
@@ -33,13 +33,20 @@ def get_build_plan(
     app: str,
     *,
     project_data: Dict[str, Any],
+    strict_platform_names: bool = False,
+    allow_app_characters: bool = False,
 ) -> Iterable[BuildInfo]:
     """Get a build plan for a given application.
 
     :param app: The name of the application (e.g. snapcraft, charmcraft, rockcraft)
     :param project_data: The raw dictionary of the project's YAML file. Normally this
         is what's output from ``yaml.safe_load()``.
+    :param strict_platform_names: Whether to strictly validate all platform names.
+    :param allow_app_characters: Whether platform name validation allows characters
+        that are reserved for use by applications.
     :returns: An iterable containing each possible BuildInfo for this file.
+    :raises: InvalidPlatformNameError if strict validation is turned on and a platform
+        name is incorrect.
 
     This function is an abstraction layer over the general build planners, taking
     the application's name and its raw data and returning an exhaustive build plan
@@ -47,6 +54,12 @@ def get_build_plan(
     in a forward-compatible manner as it adds more logic around selecting build
     planners or special behaviour for more apps.
     """
+    if strict_platform_names:
+        for name in project_data.get("platforms", {}):
+            validators.validate_strict_platform_name(
+                name, allow_app_characters=allow_app_characters
+            )
+
     planner = _APP_SPECIFIC_PLANNERS.get(app, get_platforms_build_plan)
     if app == "charmcraft":
         return planner(project_data)